+++ frontend/src/App.jsx
@@ -4,8 +4,8 @@ import Pokefight from "./pages/Pokefight";
 import Pokefeed from "./pages/Pokefeed";
 import Pokecatch from "./pages/Pokecatch";
 import Pokedex from "./pages/Pokedex";
-import Header from "./components/App components/Header";
-import Footer from "./components/App components/Footer";
+import Header from "./components/App-components/Header";
+import Footer from "./components/App-components/Footer";
 import "./App.css";
 
 function App() {
