+++ frontend/src/components/Pokedex/css/displaypokemon.css
