+++ frontend/src/components/Pokedex/css/searchbar.css
